requirements.txt
Flask==0.12.2
python-memcached==1.58
requests==2.14.2